Return type long

+++ data_structure/practical2/pr7-1.c
@@ -2,7 +2,7 @@
 
 #include <stdio.h>
 
-int factorial(int n) {
+long int factorial(int n) {
     if (n == 0) {
         return 0;
     }
